Update Makefiles to remove hardwiring for building GIGCsa
We have removed a lot of the original hardwiring in the
makefiles, ESMA_arch.mk, and ESMA_base.m files in favor of
dynamic querying.  For example, we now get the netCDF and
MPI directory paths by running the appropriate command rather
than guessing what it should be.

This configuration successfully built on Odyssey.  May need
to be synced with the more recent Hercules tag later on.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>