Updated gigc_chunk_mod.F90 for latest HEMCO version
We have updated the gigc_chunk_mod.F90 to be consistent
with the latest version of HEMCO in GEOS-Chem.  This version
gets rid of the State_Chm%TRAC_TEND array in order to
save memory.

NOTE: The new gigc_chunk_mod.F90 was taken from Christoph
Keller.  It seems that this is the only file that has
been modified lately.  We will synchronize with the more
recent GMAO tag shortly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>