Remove kludge for ifort13 in GCHP/Makefile
We have now fixed an issue in the Makefile_header.mk file.  When the
Intel Compiler is used and HPC=yes, then the compile command will now
be set to "mpifort" instead of "ifort".  This will turn on the "MPI
magic".

Seb Eastham had added a kludge to work around this, but the fix to
Makefile_header.mk is more robust.

Signed-off-by: Robert Yantosca <yantosca@clhead.dal.acenet.ca>