Implemented debug structure for MAPL
A parameter can now be set in CAP.rc which controls the amount of
information supplied by MAPL for debugging purposes. Currently this
only affects ExtData. Setting the level greater than 0 will result
in increasing amounts of data being provided about which file is
being read, what the requested sample is, which samples MAPL thinks
are available and so on.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>