Modified ExtData to accept non-constant time intervals in NetCDF data
ExtData reads the starting time and date for a file based on the
"units" attribute, and then determines when each of the file's N
samples will take place based on the time increment attribute.
However, this does not account for files with irregular timing,
or where the "units" attribute is only used as a reference time.

For example, a file with a single datapoint for each year from
1998-2010 would contain 23 samples. The "units" attribute might
be "days since 1990-01-01 00:00:00", with an approximate
increment of "365". The file's "time" co-ordinate would give the
exact timings as a vector, ie [ 2922, 3287, 3652 ... ]. Previously,
ExtData would have ignored the timing vector and placed the samples
at 1990-01-01, 1991-01-01, 1992-01-01, 1992-12-31, 1992-12-31... and
so-on, failing to get the correct start time and also responding
incorrectly to the presence of leap years.

This code has been replaced so that the time vector is read explicitly
and used to offset the date given in the "units" attribute. To
accomplish this, the GetDate routine also had to be overloaded to handle
an increment (in seconds) greater than the maximum value which can be
stored in a 32-bit integer.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>