Fixed unit conversions in gigc_chunk_mod
Previously, the units in GCHP were converted (e.g. from kg/box to
v/v dry) depending on whatever the current routine required, and then
allowed to remain in that unit until another conversion was necessary.
However, all of the unit conversion routines assume that conversions
will be performed in pairs. Example: say that we start in units of A,
but that function f() requires units of B and function g() requires
units of C. Previously, we performed the following:

A --> B
f(B)
B --> C
g(C)
C --> A

However, this resulted in the imprinting of a moisture signature. To
avoid this, we now perform all conversions with immediate reversion, ie

A --> B
f(B)
B --> A
A --> C
g(C)
C --> A

This prevents the imprinting from occurring, and conserves mass.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>