Rewrote file handling in ExtData to handle extrapolation
Files being read in by ExtData can now be "extrapolated" beyond
the years formally covered by the data. If a sample outside of
the range of the data is given, then data from the closest valid
year will be cycled until the data is within range. This implicitly
assumes that all data is defined to cover full years; partial years
of data will likely cause issues. The default "reference time" for
an entry in ExtData has also been changed to be January 1st of
either the current year minus 20 years, or of 1985 - whichever is
earlier. This prevents issues with file series which are defined
over a range which is prior to the current date where no reference
time is given, as the code previously used the current time and
data as a reference time. Debug output still needs to be trimmed down.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>