Fixed a bug in Extract_ which returned the wrong time
The routine Extract_ returns various timing variables from ESMF,
including the current date and time in various formats. When
returning them in the format nymd and nhms, this worked fine.
However, when returning them as separate integers (year, month,
day, hour, minute, second), it appears that the date/time for the
end of the simulation was mistakenly used. This resulted in the
wrong date/time being passed to (for example) HEMCO, and therefore
incorrect time zone offsets being applied when calculating emissions.
Extract_ now returns these parameters based on the current simulation
time instead. A minor bug whereby the second was used in spite of not
being returned has also been fixed, by simply setting the second to
zero.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>