Removed check for import restart file
The flag haveImpRst is checked at the start of the simulation, and the
first simulation step is skipped if the flag is not TRUE. However,
this flag does not appear to be set by any routine, meaning that the
first simulation step is always skipped. To get around this, the flag
is now always set to TRUE during initialization.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>