Fixed unit conversion problem resulting from earlier correction
Now that the units are correctly changed to v/v dry during the
GIGC initialization stage, the units are incorrect for some of
the early routines in GIGC_Chunk_Run. To correct for this,
conversion of units to kg/kg dry has been moved to the start
of the GIGC_Chunk_Run routine.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>