Now archive previous specific humidity at end of timestep
The tracer mixing ratio was rapidly increasing at each timestep
This is because it is scaled by (current SH - previous SH) to
compensate for changes in the moist air mass, but the previous
SH was not being stored (in the field State_Met%SPHU_prev). Said
field is now archived at the end of GIGC_Chunk_Run, allowing
the scaling to be applied correctly.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>