Now call SetGridFromCtr on all CPUs
While reading input.geos, the grid is initialized and lat/lon data
such as the grid point centers are set. This step is then repeated
on all other CPUs during GIGC chunk initialization. However, the
routine used during input reading on the root CPU is not appropriate
for cubed-sphere grids, even though the call made to initialize the
grid variables is valid. Therefore the call to initialize the grid
variables (Initialize_Geos_Grid) should still be called by the root
CPU during input read-in, and then the other CPUs should replicate
this call. However, all CPUs should call SetGridFromCtr, including
the root CPU. This commit should be mirrored by a change in the
main GEOS-Chem component to skip the "COMPUTE_GRID" call when
running in an ESMF or external grid environment.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>