Fix for regional files being incorrectly processed
MAPL would previously switch the two halves of a 2D input file
if the first longitude was too far from -180 degrees. This resulted
in regional files being incorrectly processed. A more sophisticated
check has been implemented which only flips the files if the prior
condition is satisfied AND the longitude range is between 355 and 365,
such that the file is almost certainly periodic in longitude.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>