MPI now chosen entirely by environment variable ESMF_COMM
Setting the environment variable ESMF_COMM to "openmpi",
"mvapich2" or "mpi" is now sufficient to set the MPI
distribution throughout the code. The variable "MPI_ROOT"
must also be set to point to the root directory of the MPI
installation (i.e. $MPI_ROOT/bin/mpirun gives the mpirun
executable, and $MPI_ROOT/lib is the path for libraries).

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>