Pressure-transport update
GCHP now carries the tracer fields internally as double-precision
arrays. This is consistent with the precision already used internally
for transport and chemistry, so the additional overhead is near-zero,
while reducing mass conservation errors by ~10 orders of magnitude.
GCHP also now uses "dry pressure" to calculate transport fluxes,
consistent with GCC. Finally the system to calculate and transfer
pressure level edges has been overhauled, such that the dynamics/
advection component has "control" over the PLE array. This requires
that the advection component is "run" even when transport is off,
although in this case no actual advection will occur. The chemistry
component has been modified to account for this. Also, to ensure
mass conservation, the mass-correction scaling factor in FV3 has been
restored to REAL*8 precision. This means that simulations with
differing numbers of cores will give different results at the machine-
precision level, but this is traded off against the aforementioned
improvement in the mass conservation error. Finally, the temperature
and specific humidity fields are now read in in the same way as in GCC,
with both X1 (before advection) and X2 (after advection) required for
the variables TMPU, SPHU and PS. As such, a change in advection time
step length must be propagated to the target read time for these
variables.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>