Fix for ExtData extrapolation edge cases
When reading in a single file, ExtData can run into an edge case
where a time slot is available but is not suitable for the requested
bracket. This can result in failure, as the current file is incorrectly
flagged as not containing the target time. This situation is now
avoided.

Signed-off-by: Sebastian David Eastham <seastham@pfe23.nas.nasa.gov>