Implemented subranges and orientation-recognition for 3D data
Files containing 3D data can now be explicitly recognized as being
"surface-oriented" (level 1 = closest to surface) if their vertical
co-ordinate variable has the attribute "positive" set to "up". The
same effect can be achieved by having the value of the level
co-ordinate be greater at layer 1 then at the final value. Files with
fewer layers than are used in the grid are also accepted, under the
assumption that this is a simple subrange of the grid's layers.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>