Fix deallocation error in ExtData
The 'end do' of a do loop was placed too early in the ExtData
finalization routine, resulting in some arrays not being correctly
deallocated. This could cause the simulation to freeze while
cleaning up. The 'end do' has been moved to the correct location
and the freeze no longer occurs.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>