Fixed issue preventing cycling of data with no year token in filename
A temporary fix has been put in place which allows data with no "%y4"
token in the filename to still cycle over a given year, by specifying
the year in the refresh template (e.g. as F2015-%m2-01T00:00:00 for
a monthly referesh of 2015 data). This gives more flexibility than the
standard "cyclic" option.

Signed-off-by: Sebastian David Eastham <seastham@pfe26.nas.nasa.gov>