Replacement of the Ad-hoc Correction for Atmospheric pressure
GCHP now handles all calculation of the pressure level edges again.
This fix required that the pressure level edges be both calculated
and stored at REAL*8 precision, rather than the REAL*4 precision
used previously. This also means that PLE and DryPLE cannot be
handled by Chem_Registry.rc, as this uses REAL*4 as the default
precision (and it is not immediately clear how to change this).

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>