Prevent user from selecting LUN 5 or 6 for log write
The default LUN for log writing is 700 in the code, but can be
overwritten from the input directory. However, if the user selects
LUN 6, this can cause problems, as units 5 and 6 are reserved for
STDIN and STDOUT. ifort seemed to be able to cope but GFortran
would die with mysterious errors. To prevent this, the code now
stops if the user tries to select either of these LUNs.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>