Day-of-week read-in now possible
Data which varies by day-of-the-week can now be read in. Such data
must be contained within a single file, and must have exactly 7 or
84 time entries. If 7 entries, these must correspond to each day of
the week, starting with Sunday. If 84, this must be 12 7-day sets,
one set for each month. Time information will be largely ignored.
This read-in mode can be selected by specifying "D" for the "Cyclic"
flag in ExtData.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>