MAPL bug fix: incorrect handling of monthly climatology in December
This update fixes a MAPL bug where the right time bracket is not
correctly handled for climatology files during December simulations.
Special climatology file handling is needed for edge months of the year
(December and January) due to the year change and month number
wrap-around. Special handling exists if the current month is January
to properly get the climatology for the left bracket (previous) month.
Similar handling should occur if the current month is December to
properly get the climatology for the right bracket (next) month.
Instead, the right bracket special handling is applied only if the
current time is January and also improperly sets the right bracket
month to December. The fix is to apply the right bracket special
handling if the current month is December and to set variable imm
to January.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>