Bug fix: allow simulation start time to precede earliest clim file time
Prior to this commit, ExtData subroutine UpdateBracketTime would exit if
the current datetime preceded the first datetime of a climatology dataset.
This occurred if starting a simulation in January prior to January 15th,
the first date in the monthly XLAI 2008 files used as climatology.

This problem is corrected by removing the forced exit and replacing it
with returning ESMF_FAILURE if both the time precedes the earliest clim
file time and no file is found. For XLAI, a file is found and thus no
failure is triggered. Updating bracket time to successfully look for the
December file proceeds later in the routine as part of the existing
handling of climatology edge months.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>