Enable HEMCO emissions diagnostics if HEMCO diagnostic .rc file is present
If a HEMCO diagnostic config file is present in the run directory and
DiagnFile is set to it in HEMCO_Config.rc, then any HEMCO emission
diagnostics listed in the file can be included in HISTORY.rc for inclusion
in history collections. HEMCO was already set up to create exports for all
diagnostics listed in the file. What is new in this commit is that the
HEMCO emissions data is now automatically copied to the export states
during CopyGCState2Exports which is invoked at the end of every dynamic
timestep within the run method in Chem_GridCompMod.

Also in this commit: minor formatting update in gigc_chunk_mod for
clarity on sources of routines used for diagnostics.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>