Abstract RATS/AERO/ANOX code to new module gigc_providerservices_mod
If GEOS-Chem is the provider of radiatively active tracers (RATS), aerosols
(AERO), or analysis OX (ANOX), then species exports are created for the
needed outputs. This code was originally throughout Chem_GridCompMod, the
primary file for the GEOS-Chem chemistry gridded component. This code is
now abstracted to a separate module to make the gridded component file
leaner. Optional outputs GCCTO3 and GCCTTO3 are also moved to the
provider module due to their dependency on PTR_O3, itself dependent on
whether GEOS-Chem is a provider to ANOX.

Also in this commit:
Clean up debug comments and commented out code in gigc_mpi_wrap and
gigc_chunk_mod.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>