Bug fix: Set State_Met%LWI from surface type fraction imports
Prior to this update State_Met%LWI was read in as a met-field import, and
regridded from lat/lon to cubed sphere in the process. This caused
incorrect land/water/ice indexes which are supposed to be 0, 1, or 2
although they are stored as real numbers. Since the source of the LWI
met-field is pre-processing using the surface fraction types from GMAO,
we now use a similar approach within GCHP to create a 0/1/2 LWI mask from
surface type fraction imports.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>