Consolidate GCHP files: remove gchp_utils.F90 and gigc_types_mod.F90
gigc_types_mod.F90 contained one line of code defining a public parameter
for the internal state prefix of GEOS-Chem species concentrations. This is
now stored in gigc_historyexports_mod.F90 because it is related to text
in HISTORY.rc. This allows removal of gigc_types_mod.F90.

gigc_utils_mod.F90 contained a few subroutines called only once. The code
for each is fairly concise and is now in the locations in which is called.
The exception is the subroutine to cap the tropopause which is now in
dao_mod.F with similar subroutines.

Also in this commit, Includes_Before_Run.H now only contains State_Met
related code. Code to set State_Chm%Species is now inline within the
GEOS-Chem chem gridded component directly following the include
of Includes_Before_Run.H. This makes the setting of State_Chm%Species
more transparent.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>