Bug fix: set BOPT=O by default
The default value of BOPT in MAPL file ESMA_base.mk was set to g,
resulting in turning on debug flags by default for certain compilers.
This manifested itself as a significant slow-down of FV3 when running
with gfortran. This bug did not impact the model when using ifort15 due
to the BOPT-dependent logic used in ESMA_base.mk.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>