Initial commit of files for use in GEOS-5, altered from C. Keller's v11-02e
NOTES - PLEASE READ!

WARNING: This code compiles within GEOS-5 when placed in Christoph
Keller's v11-02e tag cak_Icarus-1_0_GCCv1102e_003. However, testing for
a successful run is still in progress.

For now there are notes in files about what is in GEOS-5 versus what is
in GCHP. GCHP-only code is commented out. Most of the differences are
in Chem_GridCompMod.F90 (which must still be renamed and moved for GEOS-5)
and in gigc_chunk_mod.F90. Many files in GEOS-5 GIGC were deleted since
there was a previous consolidation into gigc_chunk_mod.F90 last year for
GCHP. The same consolidation works for GEOS-5.

The history exports module is still not used. The provider services
module, which was intended as a place to put RATS, AERO, and ANALYSIS_OX
code, is also not used. There have been many changes within the GEOS-5
that require additional updates to it before using it.

Overall this is a first cut set of files that are close enough to GCHP to
put them in a separate branch for GEOS-5. More work will be done to
bring the files in GCHP and GEOS closer together.

For now, sharing updates between branches will involve cherry picking
only. All commits should be very short and focused. Changes meant for
sharing should be isolated in their own commits.

To use with GEOS-5 do the following:

1. Copy Chem_GridCompMod.F90 to GEOSCHEMchem_GridCompMod.F90 into
   GEOSagcm subdirectoy GEOSCHEMchem_GridComp

2. Copy Includes_Before_Run.H and Includes_After_Run.H into GEOSagcm
   subdirectory GEOSCHEMchem_GridComp

3. Delete all existing files in GEOSagcm subdirectory
   GEOSCHEMchem_GridComp/gc_column/GIGC except for the CVS subdirectory
   (the next tag should have unnecessary files deleted to avoid this
    step in the future)

4. Copy all files in this repository branch GEOS, excluding subdirectories,
   to that directory. The files already copied to a higher level
   directory are not needed. Some of the other files may not be needed
   either (e.g. other grid comps) but they shouldn't cause a problem.

WARNING: the v11-02e tag on GEOS-5 needs an update in gc_bleeding_edge
for this code to compile. See gc_bleeding_edge commit 3ce63b67 on
3/8/2018 ("Move GCHP subroutine for capping tropopause pressure to
dao_mod.F).

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>