Set prescribed CH4 monthly mean surface distributions
This update was included in GEOS-Chem Classic v11-02e and is now
mimicked in GCHP. It was submitted by Lee Murray:

"I have prepared spatially kriged (interpolated) monthly mean surface
methane distributions from the NOAA GMD flask observations from the early
1980s through the present in a NetCDF file. This allows for seasonality
in the methane concentrations and local maxima (e.g., south central US;
western Russia) and minima. It is not appropriate to use these surface
distributions fixed throughout the entire column (spatial heterogeneity
will be reduced in the free troposphere), but I use them in all my
current simulations to fix surface concentrations and advect methane
at the cost of an additional transported tracer. Since the surface
concentration is fixed, tropospheric chemistry equilibrates much faster
than if we were to have online methane emissions, on the time-scale of
tropospheric mixing. UCX does the same thing for its treatment of
methane by default, but with a single mean surface value. Fixing surface
concentrations and allowing advection is common treatment in other
atmospheric chemistry models.

It is now generally standard practice for global models to prescribe
surface methane concentrations and allow it to advect. It doesn’t have
large impacts on the short-lived chemistry, but my maps eliminate the
artificial step gradients that the zonal bands impose and introduce
realistic and dynamic spatial gradients in the troposphere. I think
it is worth putting in tropchem as an advected tracer, but at the
very least, we should use my mean meridional gradients to impose the
fixed concentrations. The NetCDF methane files are ready and contain
observations from Feb 1983 through Dec 2016, with extrapolated trends
for 1979-2020."


Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>