Replace -lmpi_f77 with -lmpi_cxx for gfortran + openMPI
Compilation will fail when compiling with gfortran and openMPI using
newer version because -lmpi_f77 cannot be found. This library is now
deprecated and the library flag should be replaced with -lmpi_cxx. This
fix was already in place for intel compilers but not for gfortran, and is
necessary because ESMF hard-codes compile flags for different
combinations of compiler and MPI.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>