Add bash script to link GEOS-5 to GEOS-Chem git repository files
This script can be run from anywhere on Discover. It is interactive and
the user to prompted to choose which, if any, of the options to execute.
This script will change over time and generally will be version-dependent.
It is therefore tracked here along with other files used in GEOS-5.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>