Remove 2nd instance of T2M in registry; rm fjx files not used
The inclusion of T2M twice in the registry file resulted in
a compile error. Removing it corrects the issue.

Several of the fast-jx *.dat files are remants from older
versions of GEOS-Chem. They are now removed, along with
their references in the shell script to link GEOS-5 and
GEOS-Chem.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>