Remove call to Init_TOMS since already done in call to GC_Init_Extra
This update was submitted by Haipeng Lin as part of an update to
unify all modules in GeosCore to accept the same set of arguments.
This update in GCHP is necessary for compatibility with the new GEOS-Chem
implementation of Init_TOMS since that routine now takes additional
arguments. Removing the redundant call in gigc_chunk_mod.F90 avoids
the incompatibility.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>