Force usage of GCC libstdc++ rather than system (default) if using gfortran
This update was submitted by Plamen G Krastev of Harvard Research
Computing to address compilation of failure of GCHP with gfortran on
the Odyssey compute cluster at Harvard. He wrote that:

"I also needed to add '-L$GCC_HOME/lib64 -lstdc++' to the linker.
The problem is that while the ESMF library is built against the
correct version of libstdc++ which comes with GCC 7.1.0

/n/helmod/apps/centos7/Core/gcc/7.1.0-fasrc01/lib64/libstdc++.so
-> libstdc++.so.6.0.23

the executables try to link against the system libstdc++

/usr/lib64/libstdc++.so.6 -> libstdc++.so.6.0.19

which doesn't have GLIBCXX_3.4.20, etc. So one needs to explicitly link
against the correct libraries for consistency (this may be added
somewhere in the GIGC.mk file in the GCHP source code directory)."

This update update is further isolated for use only if the compiler is
gfortran.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>