Initial implementation of GCHP run directory creation from source code repo
Within GCHP, use the createRunDir.sh script. All files that are copied
to the run directory are stored in new directory Run/. Users are asked
a series of questions to set up the run directory, including path to
ExtData, path to new run directory, simulation type, and whether they
wish the run directory files (non-data) to be put under version control.
The script accepts an optional argument for run directory name. Otherwise
a default name of gchp_{simulation} is used. If the run directory already
exists then the user is prompted to enter a different name or quit.

The benchmark and standard simulations are set up by default for a 1-hr
c24 simulation starting on July 1, 2016 using 0.25x0.3125 GEOS-FP data
and 6 cores. The RnPbBe simulation is the same except that it starts on
January 1, 2016.

Once a run directory is created, users should navigate there and set up
the run directory as usual, setting the source code symbolic link and
editing runConfig.sh, environment files, run script, etc. If there is
demand to add some of these configurations during run directory creation
we can add those in the future.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>