Set Bashrc symbolic link in run directory using setBashrc script
The symbolic link is now used for compiling in build.sh and for
running with all sample run scripts. After run directory creation, users
should execute setBashrc and pass it the path to their custom environment
file for use in compiling and running.

This updates simplifies the user experience because it eliminates the
need to copy a bashrc to the run directory and edit the run script to
include the bashrc filename. User environment files can now be stored
in a single location anywhere on their network.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>