Change run directory naming to no longer assume use of bash shell
Previously the naming of the Bashrc symbolic link, bashrcSamples
subdirectory and files, and setBashrc all implicitly assumed use of
a bash shell. However, users can create their own environment files
using other shells. The naming is now changed to reflect that, with
the following updates:

 - Bashrc symbolic link is now gchp.env
 - setBashrc script is now setEnvironment
 - bashrcSamples/ subdirectory is now environmentFileSamples/
 - *.bashrc sample files are now *.env

The README is updated to reflect these changes.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>