Always pass HEMCO phase 2 to Emissions_Run for emissions calculations
Previously GCHP passed its local phase to emissions_run to calculate
emissions. This worked for GCHP phase -1 because the implementation
in HEMCO in previous versions checked only that the phase was not 1.
Recent updates to HEMCO, however, now restrict the calculation to phase
2 only resulting in all zero emissions in GCHP.

This fix removes the use of local GCHP phase in HEMCO and instead defines
and passes the HEMCO phase that corresponds to the target functionality.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>