Use HEMCO sea salt extension instead of offline sea salt emissions
There was a bug in the offline sea salt emissions settings in which
the offline equivalent of the "model sea salt Br-" was on by default in
GCHP. It is off by default in GEOS-Chem Classic and therefore should be
off by default in GCHP as well.

The fix is to comment out the BrSALA and BrSALC sea salt emissions lines
in HEMCO_Config.rc. However, this fix results in Rosenbrock integration
errors. To avoid this we are turning off the sea salt offline emissions
temporarily and using the sea salt extension instead, which has
reasonable agreement with GEOS-Chem Classic output.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>