Now deallocate the State_Diag object in GCHP simulations
We had not been deallocating the State_Diag object when running
GCHP (or GEOS-5) simulations.  This was preventing all of the timing
information from being printed out to the log file when using
gfortran and OpenMPI 3.

In the Finalize_ routine of Chem_GridCompMod.F90, we pass State_Diag
to routine GIGC_Chunk_Final (of gigc_chunk_mod.F90).

In GIGC_Chunk_Final, we call the Cleanup_State_Diag routine of
Headers/state_diag_mod.F90

NOTE: GCHP runs seem to finish normally, but then an MPI error occurs
during finalization, apparently after GEOS-Chem and the Import/Export
States have been finalized.  We do not yet know what the source
of this issue is.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>