Bug fix: use checkpoint file as restart for runs 2+ in multirun option
This update corrects a previous silent bug in which the initial restart
file specified in runConfig.sh was used as the restart when doing
multiple subsequent runs (i.e. in the presence of cap_restart). The proper
behavior when using the multirun option (runScriptSamples/multirun.sh)
is to use the checkpoint file (snapshot of end of last run) if the run
is continuing from where the last run left off.

This is now corrected by automatically updating the entry for the initial
restart file in GCHP.rc with the name of the checkpoint file if
the file cap_restart is present just prior to execution. For simplicity
the checkpoint file no longer contains the grid resolution in its name.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>