Now broadcast extra fields of Input_Opt introduced in 12.2.0
Several fields of Input_Opt changed in 12.2.0, mostly due to the passive
species and ObsPack updates.  These have now been added to the module
gigc_mpi_wrap.F90 so that they can be broadcasted to all cores.

NOTE: There may still be some issues to debug.  The code compiles
but segfaults.  Another commit might be necessary to fix this.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>