Add error trap in Chem_GridCompMod.F90 if Olson data are all zeroes
We have now added an error trap in the Run_ method of module
GCHP/Chem_GridCompMod.F90 that will stop the run if the Olson land
map data is returned as all zeroes.  This will prevent a silent bug
where the GCHP run would continue with erroneous data, and then drop
a segfault at the end of the simulation.

This is a known issue that happens when compiling GCHP with GNU Fortran,
and is probably due to a software issue deep down in ESMF or MAPL.
This issue does not occur when compiling GCHP with Intel Fortran.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>