Read input.geos on all threads rathen than on one thread with broadcast
This update reduces the hard-coded dependency between the GCHP and
GEOS-chem repositories by removing the broadcast call for every Input_Opt
variable used in GEOS-Chem. As the fields of Input_Opt change in GEOS-Chem,
GCHP will no longer require updating in all cases. This thereby removes
the risk of compile errors when removing a Input_opt variable and the
risk of introducing silents bugs when adding a new Input_Opt variable.

This update also supports moving towards a common GCHP code used in both
offline and online (GEOS-5) simulations. The cost is potential additional
time to read input.geos since it is done by all cores, but this is
negligible compared to run-time and the cost of errors introduced during
GEOS-Chem development.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>