Remove Mexico_Mask.001x001.nc file from ExtData, HEMCO_Config
The Mexico_Mask.001x001.nc file is only needed for the CAN-MEX CH4
inventory (which is not used for full-chem simulations).  None of the
default emissions inventories for fullchem or RnPbBe use this mask.
Therefore, we can remove it from both the ExtData and HEMCO_Config
template files in order to prevent GCHP from stopping with an error
in the data-read phase.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>