Update multirun option to allow resubmission using existing cap_restart
Prior to this update cap_restart was always deleting when executing
gchp.multirun.sh. This was problematic if there was a crash during a
previous multirun and one wanted to simply resubmit starting at the end
of the last good run. With this update this is now possible without any
modification to runConfig.sh or any other input files.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>