Add cmake and GNUmakefile for new MAPL
This is the last consecutive commit adding and updating files to match
files from GMAO. GCHP does not currently run with these files.
Subsequent commits will apply custom updates for GCHP to work with the
new version of MAPL. This version of MAPL corresponds to the
Integration branch of NASA repository GEOSagcm as of Nov 26, 2018.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>