Reimplement FV3 update: Restrict mass balance operation to the troposphere
This update was submitted by Seb Eastham and previously committed on
Dec 6, 2018 (commit 7a4589c2). It is now reimplemented in the updated
version of FV3.

Original commit message:
Mass balancing is now performed by scaling only tropospheric levels.
This slightly increases spurious mixing in the troposphere but
prevents it from hitting the stratosphere, where the use of fixed
pressure levels means that the rebalance should be having no effect.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>