Compilation error fix: remove // from within _ASSERT argument list
One line in GMAO_pFIO/Variable.F90 has fortran string concatenation //
within a call to _ASSERT macro. The concatenation characters are
incorrectly interpreted as C comments, thus no end to the macro was found,
triggering a compiler error. This issue is likely a result of the custom
compile flags used in GCHP. This commit includes a quick fix for this
problem.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>