Remove dependency on FLAP library for o-server to enable compilation
The FLAP external library is currently being replaced by GMAO in GEOS
with an in-house library. We therefore will not work towards acquiring
and building FLAP for use with MAPL v1.0. Disabling all dependencies is
therefore necessary to allow successful compilation.

This updates also includes turning off a couple calls to
omp_set_num_threads, which is related to use of FLAP.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>