Rename internal checkpoint file post-run to avoid MAPL crash during rerun
MAPL v1.0.0 requires that the output checkpoint file be purged prior to
reusing the run directory. To avoid deleting the file, all sample run
scripts now rename it after the run so that the next run can proceed
without issue. The new name, gcchem_internal_checkpoint.restart.nc, is
used as the restart file for runs 2+ if using the multirun option.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>