Implement reverse time stepping option; set as run-time parameter in CAP.rc
This update is to facilitate starting to implement a GCHP adjoint model.
It is based on code provided by Atanas Trayanov to run the MAPL
cap clock backwards. Added an additional feature to set time direction
at run-time rather than compile time. To run the clock backwards set
REVERSE_TIME to 1 in CAP.rc. Start time and end times correspond to
the end and start times in CAP.rc respectively.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>