Use 1-year time interval for ExtData missing file search to speed up read
If an input file with a date token is not found then a search is
performed in ExtData to look for the last file preceding the simulation
time that exists. The search begins at the reference time hard-coded in
MAPL as 1985-01-01. For the case of simulations without meteorology
available, such as simulating the future with past meteorology, the
iteration went at a frequency of 1-day, resulting in a very large
performance hit. This update changes that frequency to 1-year thereby
speeding up the model for this particular simulation scenario.

While this update primarily addresses the slow-down for the case of
using past meteorology for future simulations, it also applies to all
other instances of input read where a search for past year files is
required.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>