Add JNoonFrac field to JValuesLocalNoon collection in HISTORY templates
We now have to add the JNoonFrac field to the JValuesLocalNoon
collections in HISTORY.rc.standard and HISTORY.rc.benchmark.  This is
the fraction of times it was local noon in a grid box.  JNoonFrac is
needed to divide the local noon J-values (which are now cumulative sums
when it is local noon) in post-processing.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>