Add volcano base emission diagnostics to HISTORY.rc templates
If we are using HEMCO volcano base emissions, then we have to add the
emission fields VOLCANO_SO2_DEGAS and VOLCANO_SO2_ERUPT to the
"Emissions" collection in the files HISTORY.rc.standard and
HISTORY.rc.benchmark template files.  In GC 12.3.1, we have now
reverted to using these base emissions.

If, on the other hand, we are using the HEMCO volcano extension, then
we have to add the emission fields EmisSO2_VolcDegas and EmisSO2_VolcErupt
to the HISTORY.rc template files.  This will match the emission
diagnostic names in the HEMCO_Diagn.rc file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>